+++ src/plots/Subset/QvisSubsetPlotWindow.C
@@ -445,6 +445,9 @@ QvisSubsetPlotWindow::CreateWindowContents()
 //   Brad Whitlock, Wed Jul 20 18:10:08 PST 2005
 //   Added pointSizePixels.
 //
+//   Hank Childs, Thu Jun  8 13:41:24 PDT 2006
+//   Fix compiler warning for casting.
+//
 // ****************************************************************************
 
 void
@@ -520,7 +523,7 @@ QvisSubsetPlotWindow::UpdateWindow(bool doAll)
             break;
         case 10: // opacity
             overallOpacity->blockSignals(true);
-            overallOpacity->setValue((float)subsetAtts->GetOpacity() * 255.f);
+            overallOpacity->setValue((int)(subsetAtts->GetOpacity() * 255.f));
             overallOpacity->blockSignals(false);
             break;
         case 11: // wireframe
